Switch to per-CPU hashmaps in eBPF kernel code
Since we track values for every packet that arrives and because the eBPF
program can run on different logical CPUs, there is potential for
locking contention on shared hashmaps. Use per-CPU hashmaps to remove
this contention from the kernel code, and sum the values accross each
CPU's instance of the hashmap in the userspace code on a userspace
application tick.

Addresses #9